requirements.txt
aiogram==2.9.2
python-dotenv~=1.0.0
aiocryptopay==0.2.1
requests